Refactor App into feature components
